modularized the modals
