+++ frontend/src/components/ProfileModal/ProfileModal.js
@@ -1,28 +1,7 @@
-import React, { useContext, useRef, useState, useEffect } from 'react';
+import React, { useContext } from 'react';
 import ReactDOM from 'react-dom';
+import ModalContext from '../../context/ModalContext'; 
 import './ProfileModal.css';
-
-const ModalContext = React.createContext();
-
-export function ModalProvider({ children }) {
-  const modalRef = useRef();
-  const [value, setValue] = useState();
-
-  useEffect(() => {
-    setValue(modalRef.current);
-  }, [])
-
-  
-
-  return (
-    <>
-      <ModalContext.Provider value={value}>
-        {children}
-      </ModalContext.Provider>
-      <div ref={modalRef} />
-    </>
-  );
-}
 
 export function ProfileModal({ onClose, children }) {
   const modalNode = useContext(ModalContext);
