+++ frontend/src/context/ModalContext.js
@@ -1,0 +1,23 @@
+import React, { useRef, useState, useEffect } from 'react';
+
+const ModalContext = React.createContext();
+
+export function ModalProvider({ children }) {
+  const modalRef = useRef();
+  const [value, setValue] = useState();
+
+  useEffect(() => {
+    setValue(modalRef.current);
+  }, [])
+
+  return (
+    <>
+      <ModalContext.Provider value={value}>
+        {children}
+      </ModalContext.Provider>
+      <div ref={modalRef} />
+    </>
+  );
+}
+
+export default ModalContext;
